requirements.txt

Pillow>=8.0.0
watchdog>=2.0.0
mutagen>=1.45.0
pikepdf>=5.0.0
hachoir>=3.1.0
python-docx>=0.8.11

pillow
hachoir
mutagen
pikepdf
fastapi
uvicorn
python-multipart    
numpy
matplotlib
scipy
steganography-tools
sqlalchemy
aiofiles